refactorisation du code JS de la page de configuration du plugin côté administrateur
